% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDOC.R
\docType{data}
\name{MSfrags}
\alias{MSfrags}
\title{MSfrags}
\format{
A Character vector of a length of 4 with Peptide Sequences
}
\source{
data-raw/data.R
}
\usage{
MSfrags
}
\description{
A Collection of fragments with protein modifications that we might expect
to find from the CCR5 gene in a mass spec experiment
}
\keyword{datasets}
