% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSstrings.R
\name{parse_mods}
\alias{parse_mods}
\title{parse_mods
Utilities for parsing MS strings}
\usage{
parse_mods(seqs, format = "Spectronaut")
}
\arguments{
\item{seqs}{Character vector, Peptide sequence strings from the spectronaut output}

\item{format}{Character Value, format of the input strings}
}
\value{
A \code{protein_mods} object, which is a data.frame with the amino acid sequences and modifications
}
\description{
This function takes strings from spectronaut output,
pulls metadata on protein modifications,
and displays the data in a pleasing format.
}
